crewai-flows-crash-course/requirements.txt
# List your project dependencies below. For example:
crewai>=0.76.9
python-dotenv>=1.0.1


